+++ blogapp/src/main/java/com/blogapp/controller/CommentController.java
@@ -1,9 +1,12 @@
 package com.blogapp.controller;
 
+import com.blogapp.payload.CategoryDto;
 import com.blogapp.payload.CommentDetailsDto;
 import com.blogapp.payload.CommentDto;
 import com.blogapp.service.CommentService;
 import jakarta.validation.Valid;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -13,6 +16,8 @@ import java.util.List;
 @RestController
 @RequestMapping("api/v1/comment")
 public class CommentController {
+
+    private static final Logger logger = LoggerFactory.getLogger(CommentController.class);
 
     private CommentService commentService;
     public CommentController(CommentService commentService){
@@ -24,6 +29,7 @@ public class CommentController {
     public ResponseEntity<CommentDetailsDto> addComment(
             @Valid @RequestBody CommentDto commentDto
     ){
+        logger.info("Success! Entering Comment object: {}", commentDto);
         CommentDetailsDto comment = commentService.addComment(commentDto);
         return new ResponseEntity<>(comment, HttpStatus.CREATED);
     }
@@ -33,8 +39,15 @@ public class CommentController {
     public ResponseEntity<String> deleteComment(
             @PathVariable Long commentId
     ){
-        String deleteComment = commentService.deleteComment(commentId);
-        return new ResponseEntity<>(deleteComment, HttpStatus.OK);
+        if(commentId != null){
+            logger.info("Successfully getting comment id: {}",commentId);
+            String deleteComment = commentService.deleteComment(commentId);
+            return new ResponseEntity<>(deleteComment, HttpStatus.OK);
+        }else {
+            logger.error("Failed! Comment id is not getting: {}", commentId);
+            return new ResponseEntity<>("Comment id not getting...", HttpStatus.BAD_REQUEST);
+        }
+
     }
 
     //http://localhost:8080/api/v1/comment/updateComment/{commentId}
@@ -43,8 +56,16 @@ public class CommentController {
             @PathVariable Long commentId,
             @RequestBody CommentDto commentDto
     ){
-        CommentDetailsDto updateComment = commentService.updateComment(commentId,commentDto);
-        return new ResponseEntity<>(updateComment, HttpStatus.OK);
+        logger.info("Update Comment details: {}", commentDto);
+        CommentDetailsDto updateComment = null;
+        if(commentId != null){
+            logger.info("Success! Updated Category id is: {}",commentId);
+            updateComment = commentService.updateComment(commentId,commentDto);
+            return new ResponseEntity<>(updateComment, HttpStatus.OK);
+        }else {
+            logger.error("Failed! Comment id is not correct: {}", commentId);
+            return new ResponseEntity<>(updateComment, HttpStatus.BAD_REQUEST);
+        }
     }
 
     //http://localhost:8080/api/v1/comment/{commentId}
@@ -52,14 +73,22 @@ public class CommentController {
     public ResponseEntity<CommentDetailsDto> findByCommentId(
             @PathVariable Long commentId
     ){
-        CommentDetailsDto comment = commentService.findByCommentId(commentId);
-        return new ResponseEntity<>(comment, HttpStatus.OK);
+        CommentDetailsDto comment = null;
+        if(commentId != null) {
+            logger.info("Success! Get the comment id is: {}", commentId);
+            comment = commentService.findByCommentId(commentId);
+            return new ResponseEntity<>(comment, HttpStatus.OK);
+        }else{
+            logger.error("Failed! Get the comment id is not correct: {}", commentId);
+            return new ResponseEntity<>(comment, HttpStatus.BAD_REQUEST);
+        }
     }
 
     //http://localhost:8080/api/v1/comment
     @GetMapping
     public ResponseEntity<List<CommentDetailsDto>> getComments(){
         List<CommentDetailsDto> comments = commentService.listOfComments();
+        logger.info("Getting all the comments objects: {}",comments);
         return new ResponseEntity<>(comments, HttpStatus.OK);
     }
