+++ blogapp/src/main/java/com/blogapp/controller/PostController.java
@@ -5,6 +5,8 @@ import com.blogapp.payload.PostDetailsDto;
 import com.blogapp.payload.PostDto;
 import com.blogapp.service.PostService;
 import jakarta.validation.Valid;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -15,6 +17,9 @@ import java.util.List;
 @RestController
 @RequestMapping("api/v1/post")
 public class PostController {
+
+    private static final Logger logger = LoggerFactory.getLogger(PostController.class);
+
     private PostService postService;
     public PostController(PostService postService){
         this.postService = postService;
@@ -26,11 +31,22 @@ public class PostController {
             @Valid @RequestPart("postDto") PostDto postDto,
             @RequestPart("file") List<MultipartFile> postImages
     ){
-        if(postImages.size() <= 3){
-            PostDetailsDto post = postService.addPost(postDto,postImages);
-            return new ResponseEntity<>(post, HttpStatus.CREATED);
-        }else{
-            throw new ImagesLimitExceedException("You can upload maximum 3 images!");
+        logger.info("Entering post object: {}", postDto);
+        PostDetailsDto post = null;
+        if(postImages != null) {
+            logger.info("Success! images is coming on this URL!");
+            if (postImages.size() <= 3) {
+                logger.info("Successfully retrieved post images: {}", postImages);
+                post = postService.addPost(postDto, postImages);
+                return new ResponseEntity<>(post, HttpStatus.CREATED);
+            } else {
+                ImagesLimitExceedException e = new ImagesLimitExceedException("You can upload maximum 3 images!");
+                logger.error("Failed! Images Limit Exceed: {} : {}", e.getMessage(), e.getStackTrace());
+                throw new ImagesLimitExceedException("You can upload maximum 3 images!");
+            }
+        }else {
+            logger.error("Failed! to retrieve post images: {}",postImages);
+            return new ResponseEntity<>(post, HttpStatus.BAD_REQUEST);
         }
     }
 
@@ -39,8 +55,14 @@ public class PostController {
     public ResponseEntity<String> deletePost(
             @PathVariable Long postId
     ){
-        String deletePost = postService.deletePostDetails(postId);
-        return new ResponseEntity<>(deletePost, HttpStatus.OK);
+        if(postId != null){
+            logger.info("Successfully getting post id: {}",postId);
+            String deletePost = postService.deletePostDetails(postId);
+            return new ResponseEntity<>(deletePost, HttpStatus.OK);
+        }else{
+            logger.error("Failed! post id is not getting: {}", postId);
+            return new ResponseEntity<>("Post id not getting...", HttpStatus.BAD_REQUEST);
+        }
     }
 
     //http://localhost:8080/api/v1/post/updatePost/{postId}
@@ -50,11 +72,23 @@ public class PostController {
             @RequestPart("postDto") PostDto postDto,
             @RequestPart("file") List<MultipartFile> postImages
     ){
-        if(postImages.size() <= 3){
-            PostDetailsDto updatePost = postService.updatePost(postId,postDto,postImages);
-            return new ResponseEntity<>(updatePost, HttpStatus.OK);
+        logger.info("Update post object: {}", postDto);
+        PostDetailsDto updatePost = null;
+
+        if(postImages != null) {
+            logger.info("Success! images is coming on this URL!");
+            if (postImages.size() <= 3) {
+                logger.info("Successfully retrieved update post images: {}", postImages);
+                updatePost = postService.updatePost(postId, postDto, postImages);
+                return new ResponseEntity<>(updatePost, HttpStatus.OK);
+            } else {
+                ImagesLimitExceedException e = new ImagesLimitExceedException("You can upload maximum 3 images!");
+                logger.error("Failed! New Images Limit Exceed: {} : {}", e.getMessage(), e.getStackTrace());
+                throw new ImagesLimitExceedException("You can upload maximum 3 images!");
+            }
         }else{
-            throw new ImagesLimitExceedException("You can upload maximum 3 images!");
+            logger.error("Failed! to retrieve update post images: {}",postImages);
+            return new ResponseEntity<>(updatePost, HttpStatus.BAD_REQUEST);
         }
     }
 
@@ -63,14 +97,22 @@ public class PostController {
     public ResponseEntity<PostDetailsDto> findByPostId(
             @PathVariable Long postId
     ){
-        PostDetailsDto post = postService.findByPostId(postId);
-        return new ResponseEntity<>(post, HttpStatus.OK);
+        PostDetailsDto post = null;
+        if(postId != null){
+            logger.info("Success! Get the post id is: {}",postId);
+            post = postService.findByPostId(postId);
+            return new ResponseEntity<>(post, HttpStatus.OK);
+        }else{
+            logger.error("Failed! Get the post id is not correct: {}", postId);
+            return new ResponseEntity<>(post,HttpStatus.BAD_REQUEST);
+        }
     }
 
     //http://localhost:8080/api/v1/post
     @GetMapping
     public ResponseEntity<List<PostDetailsDto>> getPosts(){
         List<PostDetailsDto> posts = postService.listOfPosts();
+        logger.info("Getting all the posts object: {}",posts);
         return new ResponseEntity<>(posts, HttpStatus.OK);
     }
